backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
beautifulsoup4
playwright
chromadb
langchain
langchain-community
langchain-openai
python-dotenv
httpx
apscheduler
pydantic-settings
